requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-jose
passlib
python-multipart
jinja2
sqlalchemy
python-dotenv
bcrypt
httpx
apscheduler